chore: add abtesting only for new users
